extract candidate functions out for testing
