+++ apps/web/src/components/PoolCandidatesTable/helpers.tsx
@@ -12,6 +12,12 @@ import {
 } from "@gc-digital-talent/i18n";
 import { parseDateTimeUtc } from "@gc-digital-talent/date-helpers";
 import { Spoiler } from "@gc-digital-talent/ui";
+import {
+  CandidateExpiryFilter,
+  PoolStream,
+  PublishingGroup,
+} from "@gc-digital-talent/graphql";
+import { notEmpty } from "@gc-digital-talent/helpers";
 
 import {
   OrderByRelationWithColumnAggregateFunction,
@@ -29,6 +35,15 @@ import useRoutes from "~/hooks/useRoutes";
 import { getFullNameLabel } from "~/utils/nameUtils";
 
 import cells from "../Table/cells";
+import { FormValues } from "./types";
+import {
+  stringToEnumCandidateExpiry,
+  stringToEnumCandidateSuspended,
+  stringToEnumLanguage,
+  stringToEnumLocation,
+  stringToEnumOperational,
+  stringToEnumPoolCandidateStatus,
+} from "../../utils/userUtils";
 
 export const statusCell = (
   status: PoolCandidateStatus | null | undefined,
@@ -336,5 +351,109 @@ export function transformSortStateToOrderByClause(
     column: "submitted_at",
     order: SortOrder.Asc,
     user: undefined,
+  };
+}
+
+export function transformPoolCandidateSearchInputToFormValues(
+  input: PoolCandidateSearchInput | undefined,
+): FormValues {
+  return {
+    publishingGroups: input?.publishingGroups?.filter(notEmpty) ?? [],
+    classifications:
+      input?.applicantFilter?.qualifiedClassifications
+        ?.filter(notEmpty)
+        .map((c) => `${c.group}-${c.level}`) ?? [],
+    stream: input?.applicantFilter?.qualifiedStreams?.filter(notEmpty) ?? [],
+    languageAbility: input?.applicantFilter?.languageAbility ?? "",
+    workRegion:
+      input?.applicantFilter?.locationPreferences?.filter(notEmpty) ?? [],
+    operationalRequirement:
+      input?.applicantFilter?.operationalRequirements?.filter(notEmpty) ?? [],
+    equity: input?.applicantFilter?.equity
+      ? [
+          ...(input.applicantFilter.equity.hasDisability
+            ? ["hasDisability"]
+            : []),
+          ...(input.applicantFilter.equity.isIndigenous
+            ? ["isIndigenous"]
+            : []),
+          ...(input.applicantFilter.equity.isVisibleMinority
+            ? ["isVisibleMinority"]
+            : []),
+          ...(input.applicantFilter.equity.isWoman ? ["isWoman"] : []),
+        ]
+      : [],
+    pools:
+      input?.applicantFilter?.pools
+        ?.filter(notEmpty)
+        .map((poolFilter) => poolFilter.id) ?? [],
+    skills:
+      input?.applicantFilter?.skills?.filter(notEmpty).map((s) => s.id) ?? [],
+    priorityWeight: input?.priorityWeight?.map((pw) => String(pw)) ?? [],
+    poolCandidateStatus: input?.poolCandidateStatus?.filter(notEmpty) ?? [],
+    expiryStatus: input?.expiryStatus
+      ? input.expiryStatus
+      : CandidateExpiryFilter.Active,
+    suspendedStatus: input?.suspendedStatus
+      ? input.suspendedStatus
+      : CandidateSuspendedFilter.Active,
+    govEmployee: input?.isGovEmployee ? "true" : "",
+  };
+}
+
+export function transformFormValuesToFilterState(
+  data: FormValues,
+): PoolCandidateSearchInput {
+  return {
+    applicantFilter: {
+      languageAbility: data.languageAbility
+        ? stringToEnumLanguage(data.languageAbility)
+        : undefined,
+      qualifiedClassifications: data.classifications.map((classification) => {
+        const splitString = classification.split("-");
+        return { group: splitString[0], level: Number(splitString[1]) };
+      }),
+      qualifiedStreams: data.stream as PoolStream[],
+      operationalRequirements: data.operationalRequirement
+        .map((requirement) => {
+          return stringToEnumOperational(requirement);
+        })
+        .filter(notEmpty),
+      locationPreferences: data.workRegion
+        .map((region) => {
+          return stringToEnumLocation(region);
+        })
+        .filter(notEmpty),
+      equity: {
+        ...(data.equity.includes("isWoman") && { isWoman: true }),
+        ...(data.equity.includes("hasDisability") && { hasDisability: true }),
+        ...(data.equity.includes("isIndigenous") && { isIndigenous: true }),
+        ...(data.equity.includes("isVisibleMinority") && {
+          isVisibleMinority: true,
+        }),
+      },
+      pools: data.pools.map((id) => {
+        return { id };
+      }),
+      skills: data.skills.map((id) => {
+        return { id };
+      }),
+    },
+    poolCandidateStatus: data.poolCandidateStatus
+      .map((status) => {
+        return stringToEnumPoolCandidateStatus(status);
+      })
+      .filter(notEmpty),
+    priorityWeight: data.priorityWeight.map((priority) => {
+      return Number(priority);
+    }),
+    expiryStatus: data.expiryStatus
+      ? stringToEnumCandidateExpiry(data.expiryStatus)
+      : undefined,
+    suspendedStatus: data.suspendedStatus
+      ? stringToEnumCandidateSuspended(data.suspendedStatus)
+      : undefined,
+    isGovEmployee: data.govEmployee ? true : undefined, // massage from FormValue type to PoolCandidateSearchInput
+    publishingGroups: data.publishingGroups as PublishingGroup[],
   };
 }
