+++ apps/web/src/components/PoolCandidatesTable/types.ts
@@ -1,0 +1,16 @@
+export type FormValues = {
+  languageAbility: string;
+  classifications: string[];
+  stream: string[];
+  operationalRequirement: string[];
+  workRegion: string[];
+  equity: string[];
+  poolCandidateStatus: string[];
+  priorityWeight: string[];
+  pools: string[];
+  skills: string[];
+  expiryStatus: string;
+  suspendedStatus: string;
+  publishingGroups: string[];
+  govEmployee: string;
+};
